Starting to refactor
